Separated native type handling

+++ messagepax/src/org/dedee/messagepax/tests/TestMessagePaxSerializer.java
@@ -24,11 +24,11 @@ public class TestMessagePaxSerializer extends TestCase {
 
 	public void testByte() throws Exception {
 		MessagePaxSerializer s = new MessagePaxSerializer(buf);
-		s.writeByte(null);
-		s.writeByte(0);
-		s.writeByte(1);
-		s.writeByte(-32);
-		s.writeByte(-64);
+		s.writeInteger(null);
+		s.writeInteger(0);
+		s.writeInteger(1);
+		s.writeInteger(-32);
+		s.writeInteger(-64);
 		assertEquals("C00001E0D0C0", s.toHexString());
 	}
